Object of class WR

Creates an object with data and coordinate of class WR to use with mapping function or available in other R "maps" packages.

Usage

WR(data, colID = NULL,
   unit = c("country", "nato", "ocde", "continent",
            "region", "subregion", "region_wb",
            "type_income", "type_economy"),
   matchWith = c("country", "iso2", "iso3", "iso3_eh",
                 "iso3_numeric", "iso3_un", "iso2_wb",
                 "iso3_wb", "name_formal", "name_wb"),
   res = c("low", "hi"), show_wr = TRUE, subset = NULL,
   add = NULL, new_var_names = NULL,
   aggregation_fun = sum, aggregation_unit = NULL, aggregation_var = NULL,
   facets = NULL, check.unit.names = TRUE, dir = NULL, use_cache = TRUE,
   print = FALSE, use_internet = TRUE, crs = NULL)

Arguments

data

a data.frame object with variables to display

colID

character value or columns number indicating the column with unit names

unit

the type of world statistical unit

matchWith

the type of id to check:

"country"	if country names

"iso2"	if iso2 code

"iso3"	if iso3 code

"iso3_eh"	if iso3_eh code

"iso3_numeric"	if iso3 numeric code

"iso3_un"	if iso3 United Nations

"iso2_wb"	if iso2 World Bank

"iso3_wb"	if iso3 World Bank

"name_formal"	if formal names

"name_wb"	if World Bank names

res

map resolution

show_wr

logical value set to TRUE indicating if the entire map is drawn or only the coordinates linked to the input data

subset

a formula indicating the condition to subset the data, see the Details

add

a formula to add new transformed variables starting from the variables in the data

new_var_names

a character value or vector indicating the names of the new variables created in add

aggregation_fun

function to use when data are aggregated

aggregation_unit

variable name by which the unit are aggregated

aggregation_var

variable name with value to aggregate

facets

variable(s) name to split the data

check.unit.names

a logical value indicating if the colID names are checked with unit names

dir

local directory in which shape files are stored

use_cache

a logical value indicating whether use the cache

print

a logical value indicating whether print the nomatched names

use_internet

a logical value indicating wheter the coordinates are downloaded from https://github.com/mappinguniverse/geospatial. If FALSE the maps downloaded during package installation will be used

crs

coordinate reference system. Look at st_crs

Details

The function links (automatically) the id in the data and the coordinates for the given unit.

Since the names (or codes) provided in the data given in input must be checked with the unit names (or codes) available in the package, the check.unit.names provides a preliminary check.

subset provide an expression to subset the data, using a formula with the logical operators. For example, sub-samples of the data can be selected as follows: ~I("Variable 1" == "condition 1" & "Variable 2" != "condition 2") or for example, ~I("Variable 1" > "condition 1" | "Variable 2" != "condition 2").

Value

An object of class WR, with data and coordinates to use in functions which perform map.

See also

Examples

data("popWR")

wr <- WR(data = popWR, colID = "country_code",
         matchWith = "iso3_eh", res = "low")
str(wr, 1)
#> Classes ‘sf’, ‘WR’, ‘WR’ and 'data.frame':	248 obs. of  24 variables:
#>  $ country_load: chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ name_formal : chr  "Republic of Indonesia" "Malaysia" "Republic of Chile" "Plurinational State of Bolivia" ...
#>  $ name_wb     : chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ iso2        : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3        : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ country_code: chr  "idn" "mys" "chl" "bol" ...
#>  $ iso3_numeric: chr  "360" "458" "152" "068" ...
#>  $ iso3_un     : chr  "360" "458" "152" "068" ...
#>  $ iso2_wb     : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3_wb     : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ continent   : chr  "Asia" "Asia" "South America" "South America" ...
#>  $ region      : chr  "Asia" "Asia" "Americas" "Americas" ...
#>  $ subregion   : chr  "South-Eastern Asia" "South-Eastern Asia" "South America" "South America" ...
#>  $ region_wb   : chr  "East Asia & Pacific" "East Asia & Pacific" "Latin America & Caribbean" "Latin America & Caribbean" ...
#>  $ type        : chr  "Sovereign country" "Sovereign country" "Sovereign country" "Sovereign country" ...
#>  $ type_economy: chr  "Emerging" "Developing" "Emerging" "Emerging" ...
#>  $ type_income : chr  "Lower middle income" "Upper middle income" "Upper middle income" "Lower middle income" ...
#>  $ nato        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ ocde        : logi  FALSE FALSE TRUE FALSE FALSE FALSE ...
#>  $ country     : Factor w/ 265 levels "","Afghanistan",..: 112 150 45 26 193 10 56 111 46 117 ...
#>  $ total       : num  2.68e+08 3.15e+07 1.87e+07 1.14e+07 3.20e+07 ...
#>  $ male        : num  1.35e+08 1.62e+07 9.23e+06 5.70e+06 1.59e+07 ...
#>  $ female      : num  1.33e+08 1.53e+07 9.50e+06 5.65e+06 1.61e+07 ...
#>  $ geometry    :sfc_MULTIPOLYGON of length 248; first list element: List of 178
#>   ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:23] "country_load" "name_formal" "name_wb" "iso2" ...
#>  - attr(*, "unit")= chr "country"
#>  - attr(*, "colID")= chr "country_code"

wr1 <- WR(data = popWR, colID = "country_code",
          matchWith = "iso3_eh", res = "low",
          add = ~I(male/total) + I(female/total))
str(wr1)
#> Classes ‘sf’, ‘WR’, ‘WR’ and 'data.frame':	248 obs. of  26 variables:
#>  $ country_load   : chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ name_formal    : chr  "Republic of Indonesia" "Malaysia" "Republic of Chile" "Plurinational State of Bolivia" ...
#>  $ name_wb        : chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ iso2           : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3           : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ country_code   : chr  "idn" "mys" "chl" "bol" ...
#>  $ iso3_numeric   : chr  "360" "458" "152" "068" ...
#>  $ iso3_un        : chr  "360" "458" "152" "068" ...
#>  $ iso2_wb        : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3_wb        : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ continent      : chr  "Asia" "Asia" "South America" "South America" ...
#>  $ region         : chr  "Asia" "Asia" "Americas" "Americas" ...
#>  $ subregion      : chr  "South-Eastern Asia" "South-Eastern Asia" "South America" "South America" ...
#>  $ region_wb      : chr  "East Asia & Pacific" "East Asia & Pacific" "Latin America & Caribbean" "Latin America & Caribbean" ...
#>  $ type           : chr  "Sovereign country" "Sovereign country" "Sovereign country" "Sovereign country" ...
#>  $ type_economy   : chr  "Emerging" "Developing" "Emerging" "Emerging" ...
#>  $ type_income    : chr  "Lower middle income" "Upper middle income" "Upper middle income" "Lower middle income" ...
#>  $ nato           : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ ocde           : logi  FALSE FALSE TRUE FALSE FALSE FALSE ...
#>  $ country        : Factor w/ 265 levels "","Afghanistan",..: 112 150 45 26 193 10 56 111 46 117 ...
#>  $ total          : num  2.68e+08 3.15e+07 1.87e+07 1.14e+07 3.20e+07 ...
#>  $ male           : num  1.35e+08 1.62e+07 9.23e+06 5.70e+06 1.59e+07 ...
#>  $ female         : num  1.33e+08 1.53e+07 9.50e+06 5.65e+06 1.61e+07 ...
#>  $ I.male.total.  : 'AsIs' num  0.503561.... 0.514214.... 0.492729.... 0.502166.... 0.496622.... ...
#>  $ I.female.total.: 'AsIs' num  0.496438.... 0.485785.... 0.507270.... 0.497833.... 0.503377.... ...
#>  $ geometry       :sfc_MULTIPOLYGON of length 248; first list element: List of 178
#>   ..$ :List of 1
#>   .. ..$ : num [1:96, 1:2] 141 141 140 140 138 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 137 137 136 136 136 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 135 135 136 136 136 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:22, 1:2] 135 135 135 135 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 134 134 134 134 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 134 134 134 134 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 133 133 134 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 132 132 132 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:14, 1:2] 131 131 131 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 132 132 132 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:17, 1:2] 131 131 131 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 130 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:30, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:38, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:20, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:13, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 127 127 127 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:19, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 127 127 127 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 128 128 128 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:18, 1:2] 126 126 126 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 126 126 126 126 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:15, 1:2] 126 126 126 126 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:100, 1:2] 125 125 125 125 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:24, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:35, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 123 123 123 123 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:22, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:16, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:36, 1:2] 123 123 123 122 122 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:27, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   .. [list output truncated]
#>   ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:25] "country_load" "name_formal" "name_wb" "iso2" ...
#>  - attr(*, "unit")= chr "country"
#>  - attr(*, "colID")= chr "country_code"

wr2 <- WR(data = popWR, colID = "country_code",
          matchWith = "iso3_eh", res = "low",
          add = ~I(male/total) + I(female/total),
          new_var_names = c("Per_Male", "Per_Female"))
str(wr2)
#> Classes ‘sf’, ‘WR’, ‘WR’ and 'data.frame':	248 obs. of  26 variables:
#>  $ country_load: chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ name_formal : chr  "Republic of Indonesia" "Malaysia" "Republic of Chile" "Plurinational State of Bolivia" ...
#>  $ name_wb     : chr  "Indonesia" "Malaysia" "Chile" "Bolivia" ...
#>  $ iso2        : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3        : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ country_code: chr  "idn" "mys" "chl" "bol" ...
#>  $ iso3_numeric: chr  "360" "458" "152" "068" ...
#>  $ iso3_un     : chr  "360" "458" "152" "068" ...
#>  $ iso2_wb     : chr  "ID" "MY" "CL" "BO" ...
#>  $ iso3_wb     : chr  "IDN" "MYS" "CHL" "BOL" ...
#>  $ continent   : chr  "Asia" "Asia" "South America" "South America" ...
#>  $ region      : chr  "Asia" "Asia" "Americas" "Americas" ...
#>  $ subregion   : chr  "South-Eastern Asia" "South-Eastern Asia" "South America" "South America" ...
#>  $ region_wb   : chr  "East Asia & Pacific" "East Asia & Pacific" "Latin America & Caribbean" "Latin America & Caribbean" ...
#>  $ type        : chr  "Sovereign country" "Sovereign country" "Sovereign country" "Sovereign country" ...
#>  $ type_economy: chr  "Emerging" "Developing" "Emerging" "Emerging" ...
#>  $ type_income : chr  "Lower middle income" "Upper middle income" "Upper middle income" "Lower middle income" ...
#>  $ nato        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ ocde        : logi  FALSE FALSE TRUE FALSE FALSE FALSE ...
#>  $ country     : Factor w/ 265 levels "","Afghanistan",..: 112 150 45 26 193 10 56 111 46 117 ...
#>  $ total       : num  2.68e+08 3.15e+07 1.87e+07 1.14e+07 3.20e+07 ...
#>  $ male        : num  1.35e+08 1.62e+07 9.23e+06 5.70e+06 1.59e+07 ...
#>  $ female      : num  1.33e+08 1.53e+07 9.50e+06 5.65e+06 1.61e+07 ...
#>  $ Per_Male    : 'AsIs' num  0.503561.... 0.514214.... 0.492729.... 0.502166.... 0.496622.... ...
#>  $ Per_Female  : 'AsIs' num  0.496438.... 0.485785.... 0.507270.... 0.497833.... 0.503377.... ...
#>  $ geometry    :sfc_MULTIPOLYGON of length 248; first list element: List of 178
#>   ..$ :List of 1
#>   .. ..$ : num [1:96, 1:2] 141 141 140 140 138 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 137 137 136 136 136 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 135 135 136 136 136 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 135 135 135 135 135 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:22, 1:2] 135 135 135 135 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 134 134 134 134 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 134 134 134 134 134 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 133 133 134 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 133 133 133 133 133 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 132 132 132 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:14, 1:2] 131 131 131 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 132 132 132 132 132 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:17, 1:2] 131 131 131 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 131 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 130 131 131 131 131 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:30, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 130 130 130 130 130 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 130 130 130 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:38, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 129 129 129 129 129 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:20, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:13, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 127 127 127 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:19, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 128 128 128 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 127 127 127 128 128 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 128 128 128 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:18, 1:2] 126 126 126 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:11, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 126 126 126 126 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 127 127 127 127 127 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:15, 1:2] 126 126 126 126 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:8, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 126 126 126 126 126 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:6, 1:2] 125 125 125 125 125 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:100, 1:2] 125 125 125 125 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:24, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:35, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:10, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:12, 1:2] 123 123 123 123 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:7, 1:2] 124 124 124 124 124 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:22, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:16, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:36, 1:2] 123 123 123 122 122 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:9, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:27, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   ..$ :List of 1
#>   .. ..$ : num [1:5, 1:2] 123 123 123 123 123 ...
#>   .. [list output truncated]
#>   ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:25] "country_load" "name_formal" "name_wb" "iso2" ...
#>  - attr(*, "unit")= chr "country"
#>  - attr(*, "colID")= chr "country_code"