European names

Retrieves European statistical unit names.

Usage

getNamesEU(year = c("2021","2016", "2013", "2010", "2006", "2003"),
           unit = c("nuts0", "nuts1", "nuts2", "nuts3"), id = FALSE, all_levels = TRUE)

Arguments

year

year of the analysis

unit

the type of statistical unit

id

boolean value indicating whether the ids are returned instead of names

all_levels

a logical value indicating if all levels are returned or only the unit names

Value

A character vector or a data frame with unit names and corresponding associated levels.

See also

Examples

getNamesEU()
#>                                                 country iso2 iso3 country_code
#> 1                                               Germany   DE  DEU          276
#> 2                                               Czechia   CZ  CZE          203
#> 3                                              Bulgaria   BG  BGR          100
#> 4                                           Switzerland   CH  CHE          756
#> 5                                               Albania   AL  ALB            8
#> 6                                               Austria   AT  AUT           40
#> 7                                                Cyprus   CY  CYP          196
#> 8                                                Greece   GR  GRC          300
#> 9                                               Belgium   BE  BEL           56
#> 10                                               France   FR  FRA          250
#> 11                                              Denmark   DK  DNK          208
#> 12                                              Estonia   EE  EST          233
#> 13                                                Spain   ES  ESP          724
#> 14                                              Finland   FI  FIN          246
#> 15                                               Norway   NO  NOR          578
#> 16                                               Sweden   SE  SWE          752
#> 17                                             Slovenia   SI  SVN          705
#> 18                                          Netherlands   NL  NLD          528
#> 19                                                Italy   IT  ITA          380
#> 20                                            Lithuania   LT  LTU          440
#> 21                                           Luxembourg   LU  LUX          442
#> 22                                               Latvia   LV  LVA          428
#> 23                                           Montenegro   ME  MNE          499
#> 24                                      North Macedonia   MK  MKD          807
#> 25                                                Malta   MT  MLT          470
#> 26                                              Romania   RO  ROU          642
#> 27                                               Serbia   RS  SRB          688
#> 28                                              Croatia   HR  HRV          191
#> 29                                             Slovakia   SK  SVK          703
#> 30                                        Liechtenstein   LI  LIE          438
#> 31                                             Portugal   PT  PRT          620
#> 32                                              Hungary   HU  HUN          348
#> 33                                              Ireland   IE  IRL          372
#> 34                                              Iceland   IS  ISL          352
#> 35 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 36                                               Poland   PL  POL          616
#> 37                                               Turkey   TR  TUR          792
#>    nuts0_id                                                nuts0
#> 1        DE                                              Germany
#> 2        CZ                                              Czechia
#> 3        BG                                             Bulgaria
#> 4        CH                                          Switzerland
#> 5        AL                                              Albania
#> 6        AT                                              Austria
#> 7        CY                                               Cyprus
#> 8        EL                                               Greece
#> 9        BE                                              Belgium
#> 10       FR                                               France
#> 11       DK                                              Denmark
#> 12       EE                                              Estonia
#> 13       ES                                                Spain
#> 14       FI                                              Finland
#> 15       NO                                               Norway
#> 16       SE                                               Sweden
#> 17       SI                                             Slovenia
#> 18       NL                                          Netherlands
#> 19       IT                                                Italy
#> 20       LT                                            Lithuania
#> 21       LU                                           Luxembourg
#> 22       LV                                               Latvia
#> 23       ME                                           Montenegro
#> 24       MK                                      North Macedonia
#> 25       MT                                                Malta
#> 26       RO                                              Romania
#> 27       RS                                               Serbia
#> 28       HR                                              Croatia
#> 29       SK                                             Slovakia
#> 30       LI                                        Liechtenstein
#> 31       PT                                             Portugal
#> 32       HU                                              Hungary
#> 33       IE                                              Ireland
#> 34       IS                                              Iceland
#> 35       UK United Kingdom of Great Britain and Northern Ireland
#> 36       PL                                               Poland
#> 37       TR                                               Turkey

getNamesEU(unit = "nuts1")
#>                                                  country iso2 iso3 country_code
#> 1                                                Germany   DE  DEU          276
#> 2                                                Czechia   CZ  CZE          203
#> 3                                                Germany   DE  DEU          276
#> 4                                               Bulgaria   BG  BGR          100
#> 5                                               Bulgaria   BG  BGR          100
#> 6                                            Switzerland   CH  CHE          756
#> 7                                                Albania   AL  ALB            8
#> 8                                                Austria   AT  AUT           40
#> 9                                                 Cyprus   CY  CYP          196
#> 10                                                Greece   GR  GRC          300
#> 11                                                Greece   GR  GRC          300
#> 12                                               Austria   AT  AUT           40
#> 13                                               Austria   AT  AUT           40
#> 14                                               Belgium   BE  BEL           56
#> 15                                               Belgium   BE  BEL           56
#> 16                                               Belgium   BE  BEL           56
#> 17                                               Germany   DE  DEU          276
#> 18                                               Germany   DE  DEU          276
#> 19                                                France   FR  FRA          250
#> 20                                                Greece   GR  GRC          300
#> 21                                               Denmark   DK  DNK          208
#> 22                                               Estonia   EE  EST          233
#> 23                                               Germany   DE  DEU          276
#> 24                                               Germany   DE  DEU          276
#> 25                                               Germany   DE  DEU          276
#> 26                                                France   FR  FRA          250
#> 27                                                France   FR  FRA          250
#> 28                                                France   FR  FRA          250
#> 29                                                Greece   GR  GRC          300
#> 30                                               Germany   DE  DEU          276
#> 31                                                France   FR  FRA          250
#> 32                                               Germany   DE  DEU          276
#> 33                                               Germany   DE  DEU          276
#> 34                                                 Spain   ES  ESP          724
#> 35                                               Finland   FI  FIN          246
#> 36                                               Germany   DE  DEU          276
#> 37                                               Germany   DE  DEU          276
#> 38                                               Germany   DE  DEU          276
#> 39                                               Germany   DE  DEU          276
#> 40                                               Germany   DE  DEU          276
#> 41                                               Germany   DE  DEU          276
#> 42                                                France   FR  FRA          250
#> 43                                                France   FR  FRA          250
#> 44                                                 Spain   ES  ESP          724
#> 45                                               Finland   FI  FIN          246
#> 46                                                France   FR  FRA          250
#> 47                                                Norway   NO  NOR          578
#> 48                                                 Spain   ES  ESP          724
#> 49                                                 Spain   ES  ESP          724
#> 50                                                 Spain   ES  ESP          724
#> 51                                                 Spain   ES  ESP          724
#> 52                                                 Spain   ES  ESP          724
#> 53                                                Sweden   SE  SWE          752
#> 54                                              Slovenia   SI  SVN          705
#> 55                                           Netherlands   NL  NLD          528
#> 56                                           Netherlands   NL  NLD          528
#> 57                                           Netherlands   NL  NLD          528
#> 58                                           Netherlands   NL  NLD          528
#> 59                                                 Italy   IT  ITA          380
#> 60                                                France   FR  FRA          250
#> 61                                                France   FR  FRA          250
#> 62                                                France   FR  FRA          250
#> 63                                                France   FR  FRA          250
#> 64                                                France   FR  FRA          250
#> 65                                             Lithuania   LT  LTU          440
#> 66                                            Luxembourg   LU  LUX          442
#> 67                                                Latvia   LV  LVA          428
#> 68                                            Montenegro   ME  MNE          499
#> 69                                       North Macedonia   MK  MKD          807
#> 70                                                 Malta   MT  MLT          470
#> 71                                               Romania   RO  ROU          642
#> 72                                               Romania   RO  ROU          642
#> 73                                                Serbia   RS  SRB          688
#> 74                                                Serbia   RS  SRB          688
#> 75                                               Croatia   HR  HRV          191
#> 76                                                 Italy   IT  ITA          380
#> 77                                              Slovakia   SK  SVK          703
#> 78                                                 Italy   IT  ITA          380
#> 79                                                 Italy   IT  ITA          380
#> 80                                                 Italy   IT  ITA          380
#> 81                                         Liechtenstein   LI  LIE          438
#> 82                                                Sweden   SE  SWE          752
#> 83                                              Portugal   PT  PRT          620
#> 84                                               Hungary   HU  HUN          348
#> 85                                               Hungary   HU  HUN          348
#> 86                                               Hungary   HU  HUN          348
#> 87                                               Ireland   IE  IRL          372
#> 88                                               Iceland   IS  ISL          352
#> 89  United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 90  United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 91                                                Poland   PL  POL          616
#> 92                                               Romania   RO  ROU          642
#> 93                                               Romania   RO  ROU          642
#> 94  United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 95  United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 96                                                Poland   PL  POL          616
#> 97                                                Poland   PL  POL          616
#> 98                                                Poland   PL  POL          616
#> 99                                                Poland   PL  POL          616
#> 100                                               Poland   PL  POL          616
#> 101                                               Poland   PL  POL          616
#> 102 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 103 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 104                                             Portugal   PT  PRT          620
#> 105                                             Portugal   PT  PRT          620
#> 106                                               Turkey   TR  TUR          792
#> 107                                               Turkey   TR  TUR          792
#> 108                                               Turkey   TR  TUR          792
#> 109                                               Turkey   TR  TUR          792
#> 110                                               Turkey   TR  TUR          792
#> 111 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 112 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 113                                               Sweden   SE  SWE          752
#> 114 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 115 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 116                                               Turkey   TR  TUR          792
#> 117                                               Turkey   TR  TUR          792
#> 118                                               Turkey   TR  TUR          792
#> 119                                               Turkey   TR  TUR          792
#> 120                                               Turkey   TR  TUR          792
#> 121                                               Turkey   TR  TUR          792
#> 122 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 123 United Kingdom of Great Britain and Northern Ireland   GB  GBR          826
#> 124                                               Turkey   TR  TUR          792
#>     nuts0_id nuts1_id                                                nuts0
#> 1         DE      DE2                                              Germany
#> 2         CZ      CZ0                                              Czechia
#> 3         DE      DE1                                              Germany
#> 4         BG      BG3                                             Bulgaria
#> 5         BG      BG4                                             Bulgaria
#> 6         CH      CH0                                          Switzerland
#> 7         AL      AL0                                              Albania
#> 8         AT      AT1                                              Austria
#> 9         CY      CY0                                               Cyprus
#> 10        EL      EL4                                               Greece
#> 11        EL      EL5                                               Greece
#> 12        AT      AT2                                              Austria
#> 13        AT      AT3                                              Austria
#> 14        BE      BE1                                              Belgium
#> 15        BE      BE2                                              Belgium
#> 16        BE      BE3                                              Belgium
#> 17        DE      DED                                              Germany
#> 18        DE      DEE                                              Germany
#> 19        FR      FRJ                                               France
#> 20        EL      EL6                                               Greece
#> 21        DK      DK0                                              Denmark
#> 22        EE      EE0                                              Estonia
#> 23        DE      DEA                                              Germany
#> 24        DE      DEF                                              Germany
#> 25        DE      DEG                                              Germany
#> 26        FR      FRB                                               France
#> 27        FR      FRC                                               France
#> 28        FR      FRD                                               France
#> 29        EL      EL3                                               Greece
#> 30        DE      DE9                                              Germany
#> 31        FR      FRK                                               France
#> 32        DE      DEB                                              Germany
#> 33        DE      DEC                                              Germany
#> 34        ES      ES7                                                Spain
#> 35        FI      FI1                                              Finland
#> 36        DE      DE3                                              Germany
#> 37        DE      DE4                                              Germany
#> 38        DE      DE5                                              Germany
#> 39        DE      DE6                                              Germany
#> 40        DE      DE7                                              Germany
#> 41        DE      DE8                                              Germany
#> 42        FR      FRL                                               France
#> 43        FR      FRM                                               France
#> 44        ES      ES1                                                Spain
#> 45        FI      FI2                                              Finland
#> 46        FR      FR1                                               France
#> 47        NO      NO0                                               Norway
#> 48        ES      ES2                                                Spain
#> 49        ES      ES3                                                Spain
#> 50        ES      ES4                                                Spain
#> 51        ES      ES5                                                Spain
#> 52        ES      ES6                                                Spain
#> 53        SE      SE3                                               Sweden
#> 54        SI      SI0                                             Slovenia
#> 55        NL      NL1                                          Netherlands
#> 56        NL      NL2                                          Netherlands
#> 57        NL      NL3                                          Netherlands
#> 58        NL      NL4                                          Netherlands
#> 59        IT      ITF                                                Italy
#> 60        FR      FRE                                               France
#> 61        FR      FRF                                               France
#> 62        FR      FRG                                               France
#> 63        FR      FRH                                               France
#> 64        FR      FRI                                               France
#> 65        LT      LT0                                            Lithuania
#> 66        LU      LU0                                           Luxembourg
#> 67        LV      LV0                                               Latvia
#> 68        ME      ME0                                           Montenegro
#> 69        MK      MK0                                      North Macedonia
#> 70        MT      MT0                                                Malta
#> 71        RO      RO3                                              Romania
#> 72        RO      RO4                                              Romania
#> 73        RS      RS1                                               Serbia
#> 74        RS      RS2                                               Serbia
#> 75        HR      HR0                                              Croatia
#> 76        IT      ITC                                                Italy
#> 77        SK      SK0                                             Slovakia
#> 78        IT      ITG                                                Italy
#> 79        IT      ITH                                                Italy
#> 80        IT      ITI                                                Italy
#> 81        LI      LI0                                        Liechtenstein
#> 82        SE      SE2                                               Sweden
#> 83        PT      PT1                                             Portugal
#> 84        HU      HU1                                              Hungary
#> 85        HU      HU2                                              Hungary
#> 86        HU      HU3                                              Hungary
#> 87        IE      IE0                                              Ireland
#> 88        IS      IS0                                              Iceland
#> 89        UK      UKJ United Kingdom of Great Britain and Northern Ireland
#> 90        UK      UKK United Kingdom of Great Britain and Northern Ireland
#> 91        PL      PL2                                               Poland
#> 92        RO      RO1                                              Romania
#> 93        RO      RO2                                              Romania
#> 94        UK      UKN United Kingdom of Great Britain and Northern Ireland
#> 95        UK      UKL United Kingdom of Great Britain and Northern Ireland
#> 96        PL      PL4                                               Poland
#> 97        PL      PL5                                               Poland
#> 98        PL      PL6                                               Poland
#> 99        PL      PL7                                               Poland
#> 100       PL      PL8                                               Poland
#> 101       PL      PL9                                               Poland
#> 102       UK      UKC United Kingdom of Great Britain and Northern Ireland
#> 103       UK      UKD United Kingdom of Great Britain and Northern Ireland
#> 104       PT      PT2                                             Portugal
#> 105       PT      PT3                                             Portugal
#> 106       TR      TR2                                               Turkey
#> 107       TR      TR3                                               Turkey
#> 108       TR      TR4                                               Turkey
#> 109       TR      TR5                                               Turkey
#> 110       TR      TR6                                               Turkey
#> 111       UK      UKG United Kingdom of Great Britain and Northern Ireland
#> 112       UK      UKH United Kingdom of Great Britain and Northern Ireland
#> 113       SE      SE1                                               Sweden
#> 114       UK      UKI United Kingdom of Great Britain and Northern Ireland
#> 115       UK      UKM United Kingdom of Great Britain and Northern Ireland
#> 116       TR      TR7                                               Turkey
#> 117       TR      TR8                                               Turkey
#> 118       TR      TR9                                               Turkey
#> 119       TR      TRA                                               Turkey
#> 120       TR      TRB                                               Turkey
#> 121       TR      TRC                                               Turkey
#> 122       UK      UKE United Kingdom of Great Britain and Northern Ireland
#> 123       UK      UKF United Kingdom of Great Britain and Northern Ireland
#> 124       TR      TR1                                               Turkey
#>                                                           nuts1
#> 1                                                        Bayern
#> 2                                                         Česko
#> 3                                             Baden-Württemberg
#> 4                                 Северна и Югоизточна България
#> 5                          Югозападна и Южна централна България
#> 6                                       Schweiz/Suisse/Svizzera
#> 7                                                     Shqipëria
#> 8                                                 Ostösterreich
#> 9                                                        Κύπρος
#> 10                                         Νησιά Αιγαίου, Κρήτη
#> 11                                                Βόρεια Ελλάδα
#> 12                                                Südösterreich
#> 13                                               Westösterreich
#> 14  Région de Bruxelles-Capitale/Brussels Hoofdstedelijk Gewest
#> 15                                                Vlaams Gewest
#> 16                                              Région wallonne
#> 17                                                      Sachsen
#> 18                                               Sachsen-Anhalt
#> 19                                                    Occitanie
#> 20                                              Κεντρική Ελλάδα
#> 21                                                      Danmark
#> 22                                                        Eesti
#> 23                                          Nordrhein-Westfalen
#> 24                                           Schleswig-Holstein
#> 25                                                    Thüringen
#> 26                                        Centre — Val de Loire
#> 27                                      Bourgogne-Franche-Comté
#> 28                                                    Normandie
#> 29                                                       Αττική
#> 30                                                Niedersachsen
#> 31                                         Auvergne-Rhône-Alpes
#> 32                                              Rheinland-Pfalz
#> 33                                                     Saarland
#> 34                                                     Canarias
#> 35                                                 Manner-Suomi
#> 36                                                       Berlin
#> 37                                                  Brandenburg
#> 38                                                       Bremen
#> 39                                                      Hamburg
#> 40                                                       Hessen
#> 41                                       Mecklenburg-Vorpommern
#> 42                                   Provence-Alpes-Côte d’Azur
#> 43                                                        Corse
#> 44                                                     Noroeste
#> 45                                                        Åland
#> 46                                                Ile-de-France
#> 47                                                        Norge
#> 48                                                      Noreste
#> 49                                          Comunidad de Madrid
#> 50                                                  Centro (ES)
#> 51                                                         Este
#> 52                                                          Sur
#> 53                                                Norra Sverige
#> 54                                                    Slovenija
#> 55                                              Noord-Nederland
#> 56                                               Oost-Nederland
#> 57                                               West-Nederland
#> 58                                               Zuid-Nederland
#> 59                                                          Sud
#> 60                                              Hauts-de-France
#> 61                                                    Grand Est
#> 62                                             Pays de la Loire
#> 63                                                     Bretagne
#> 64                                           Nouvelle-Aquitaine
#> 65                                                      Lietuva
#> 66                                                   Luxembourg
#> 67                                                      Latvija
#> 68                                                    Црна Гора
#> 69                                           Северна Македонија
#> 70                                                        Malta
#> 71                                           Macroregiunea Trei
#> 72                                          Macroregiunea Patru
#> 73                                               Србија - север
#> 74                                                 Србија - југ
#> 75                                                     Hrvatska
#> 76                                                   Nord-Ovest
#> 77                                                    Slovensko
#> 78                                                        Isole
#> 79                                                     Nord-Est
#> 80                                                  Centro (IT)
#> 81                                                Liechtenstein
#> 82                                                Södra Sverige
#> 83                                                   Continente
#> 84                                           Közép-Magyarország
#> 85                                                     Dunántúl
#> 86                                              Alföld és Észak
#> 87                                                      Ireland
#> 88                                                       Ísland
#> 89                                         South East (England)
#> 90                                         South West (England)
#> 91                                       Makroregion południowy
#> 92                                            Macroregiunea Unu
#> 93                                            Macroregiunea Doi
#> 94                                             Northern Ireland
#> 95                                                        Wales
#> 96                                Makroregion północno-zachodni
#> 97                              Makroregion południowo-zachodni
#> 98                                         Makroregion północny
#> 99                                        Makroregion centralny
#> 100                                        Makroregion wschodni
#> 101                         Makroregion województwo mazowieckie
#> 102                                        North East (England)
#> 103                                        North West (England)
#> 104                                  Região Autónoma dos Açores
#> 105                                  Região Autónoma da Madeira
#> 106                                                Batı Marmara
#> 107                                                         Ege
#> 108                                                Doğu Marmara
#> 109                                                Batı Anadolu
#> 110                                                     Akdeniz
#> 111                                     West Midlands (England)
#> 112                                             East of England
#> 113                                               Östra Sverige
#> 114                                                      London
#> 115                                                    Scotland
#> 116                                                Orta Anadolu
#> 117                                              Batı Karadeniz
#> 118                                              Doğu Karadeniz
#> 119                                           Kuzeydoğu Anadolu
#> 120                                            Ortadoğu Anadolu
#> 121                                           Güneydoğu Anadolu
#> 122                                    Yorkshire and the Humber
#> 123                                     East Midlands (England)
#> 124                                                    İstanbul
getNamesEU(unit = "nuts1", all_levels = FALSE, id = FALSE)
#>   [1] "Bayern"                                                     
#>   [2] "Česko"                                                      
#>   [3] "Baden-Württemberg"                                          
#>   [4] "Северна и Югоизточна България"                              
#>   [5] "Югозападна и Южна централна България"                       
#>   [6] "Schweiz/Suisse/Svizzera"                                    
#>   [7] "Shqipëria"                                                  
#>   [8] "Ostösterreich"                                              
#>   [9] "Κύπρος"                                                     
#>  [10] "Νησιά Αιγαίου, Κρήτη"                                       
#>  [11] "Βόρεια Ελλάδα"                                              
#>  [12] "Südösterreich"                                              
#>  [13] "Westösterreich"                                             
#>  [14] "Région de Bruxelles-Capitale/Brussels Hoofdstedelijk Gewest"
#>  [15] "Vlaams Gewest"                                              
#>  [16] "Région wallonne"                                            
#>  [17] "Sachsen"                                                    
#>  [18] "Sachsen-Anhalt"                                             
#>  [19] "Occitanie"                                                  
#>  [20] "Κεντρική Ελλάδα"                                            
#>  [21] "Danmark"                                                    
#>  [22] "Eesti"                                                      
#>  [23] "Nordrhein-Westfalen"                                        
#>  [24] "Schleswig-Holstein"                                         
#>  [25] "Thüringen"                                                  
#>  [26] "Centre — Val de Loire"                                      
#>  [27] "Bourgogne-Franche-Comté"                                    
#>  [28] "Normandie"                                                  
#>  [29] "Αττική"                                                     
#>  [30] "Niedersachsen"                                              
#>  [31] "Auvergne-Rhône-Alpes"                                       
#>  [32] "Rheinland-Pfalz"                                            
#>  [33] "Saarland"                                                   
#>  [34] "Canarias"                                                   
#>  [35] "Manner-Suomi"                                               
#>  [36] "Berlin"                                                     
#>  [37] "Brandenburg"                                                
#>  [38] "Bremen"                                                     
#>  [39] "Hamburg"                                                    
#>  [40] "Hessen"                                                     
#>  [41] "Mecklenburg-Vorpommern"                                     
#>  [42] "Provence-Alpes-Côte d’Azur"                                 
#>  [43] "Corse"                                                      
#>  [44] "Noroeste"                                                   
#>  [45] "Åland"                                                      
#>  [46] "Ile-de-France"                                              
#>  [47] "Norge"                                                      
#>  [48] "Noreste"                                                    
#>  [49] "Comunidad de Madrid"                                        
#>  [50] "Centro (ES)"                                                
#>  [51] "Este"                                                       
#>  [52] "Sur"                                                        
#>  [53] "Norra Sverige"                                              
#>  [54] "Slovenija"                                                  
#>  [55] "Noord-Nederland"                                            
#>  [56] "Oost-Nederland"                                             
#>  [57] "West-Nederland"                                             
#>  [58] "Zuid-Nederland"                                             
#>  [59] "Sud"                                                        
#>  [60] "Hauts-de-France"                                            
#>  [61] "Grand Est"                                                  
#>  [62] "Pays de la Loire"                                           
#>  [63] "Bretagne"                                                   
#>  [64] "Nouvelle-Aquitaine"                                         
#>  [65] "Lietuva"                                                    
#>  [66] "Luxembourg"                                                 
#>  [67] "Latvija"                                                    
#>  [68] "Црна Гора"                                                  
#>  [69] "Северна Македонија"                                         
#>  [70] "Malta"                                                      
#>  [71] "Macroregiunea Trei"                                         
#>  [72] "Macroregiunea Patru"                                        
#>  [73] "Србија - север"                                             
#>  [74] "Србија - југ"                                               
#>  [75] "Hrvatska"                                                   
#>  [76] "Nord-Ovest"                                                 
#>  [77] "Slovensko"                                                  
#>  [78] "Isole"                                                      
#>  [79] "Nord-Est"                                                   
#>  [80] "Centro (IT)"                                                
#>  [81] "Liechtenstein"                                              
#>  [82] "Södra Sverige"                                              
#>  [83] "Continente"                                                 
#>  [84] "Közép-Magyarország"                                         
#>  [85] "Dunántúl"                                                   
#>  [86] "Alföld és Észak"                                            
#>  [87] "Ireland"                                                    
#>  [88] "Ísland"                                                     
#>  [89] "South East (England)"                                       
#>  [90] "South West (England)"                                       
#>  [91] "Makroregion południowy"                                     
#>  [92] "Macroregiunea Unu"                                          
#>  [93] "Macroregiunea Doi"                                          
#>  [94] "Northern Ireland"                                           
#>  [95] "Wales"                                                      
#>  [96] "Makroregion północno-zachodni"                              
#>  [97] "Makroregion południowo-zachodni"                            
#>  [98] "Makroregion północny"                                       
#>  [99] "Makroregion centralny"                                      
#> [100] "Makroregion wschodni"                                       
#> [101] "Makroregion województwo mazowieckie"                        
#> [102] "North East (England)"                                       
#> [103] "North West (England)"                                       
#> [104] "Região Autónoma dos Açores"                                 
#> [105] "Região Autónoma da Madeira"                                 
#> [106] "Batı Marmara"                                               
#> [107] "Ege"                                                        
#> [108] "Doğu Marmara"                                               
#> [109] "Batı Anadolu"                                               
#> [110] "Akdeniz"                                                    
#> [111] "West Midlands (England)"                                    
#> [112] "East of England"                                            
#> [113] "Östra Sverige"                                              
#> [114] "London"                                                     
#> [115] "Scotland"                                                   
#> [116] "Orta Anadolu"                                               
#> [117] "Batı Karadeniz"                                             
#> [118] "Doğu Karadeniz"                                             
#> [119] "Kuzeydoğu Anadolu"                                          
#> [120] "Ortadoğu Anadolu"                                           
#> [121] "Güneydoğu Anadolu"                                          
#> [122] "Yorkshire and the Humber"                                   
#> [123] "East Midlands (England)"                                    
#> [124] "İstanbul"                                                   
getNamesEU(unit = "nuts1", all_levels = FALSE, id = TRUE)
#>   [1] "DE2" "CZ0" "DE1" "BG3" "BG4" "CH0" "AL0" "AT1" "CY0" "EL4" "EL5" "AT2"
#>  [13] "AT3" "BE1" "BE2" "BE3" "DED" "DEE" "FRJ" "EL6" "DK0" "EE0" "DEA" "DEF"
#>  [25] "DEG" "FRB" "FRC" "FRD" "EL3" "DE9" "FRK" "DEB" "DEC" "ES7" "FI1" "DE3"
#>  [37] "DE4" "DE5" "DE6" "DE7" "DE8" "FRL" "FRM" "ES1" "FI2" "FR1" "NO0" "ES2"
#>  [49] "ES3" "ES4" "ES5" "ES6" "SE3" "SI0" "NL1" "NL2" "NL3" "NL4" "ITF" "FRE"
#>  [61] "FRF" "FRG" "FRH" "FRI" "LT0" "LU0" "LV0" "ME0" "MK0" "MT0" "RO3" "RO4"
#>  [73] "RS1" "RS2" "HR0" "ITC" "SK0" "ITG" "ITH" "ITI" "LI0" "SE2" "PT1" "HU1"
#>  [85] "HU2" "HU3" "IE0" "IS0" "UKJ" "UKK" "PL2" "RO1" "RO2" "UKN" "UKL" "PL4"
#>  [97] "PL5" "PL6" "PL7" "PL8" "PL9" "UKC" "UKD" "PT2" "PT3" "TR2" "TR3" "TR4"
#> [109] "TR5" "TR6" "UKG" "UKH" "SE1" "UKI" "UKM" "TR7" "TR8" "TR9" "TRA" "TRB"
#> [121] "TRC" "UKE" "UKF" "TR1"