Italian names

Retrieves Italian statistical unit names.

Usage

getNamesIT(year = c("2021","2020", "2019", "2018", "2017"),
           unit = c("ripartizione", "regione", "provincia", "comune"), all_levels = TRUE)

Arguments

year

year of the analysis

unit

the type of Italian statistical unit

all_levels

a logical value indicating if all levels are returned or only the unit names

Value

A character vector or a data frame with unit names and corresponding associated levels.

See also

Examples

getNamesIT()
#>   ripartizione code_ripartizione
#> 1   Nord-ovest                 1
#> 2          Sud                 4
#> 3        Isole                 5
#> 4     Nord-est                 2
#> 5       Centro                 3
getNamesIT(unit = "provincia")
#>     ripartizione                      regione                    provincia code
#> 1     Nord-ovest                     Piemonte                       Torino   TO
#> 2     Nord-ovest                     Piemonte                     Vercelli   VC
#> 3     Nord-ovest                     Piemonte                       Novara   NO
#> 4     Nord-ovest                     Piemonte                        Cuneo   CN
#> 5     Nord-ovest                     Piemonte                         Asti   AT
#> 6     Nord-ovest                     Piemonte                  Alessandria   AL
#> 7     Nord-ovest Valle d'Aosta/Vallée d'Aoste Valle d'Aosta/Vallée d'Aoste   AO
#> 8            Sud                       Molise                   Campobasso   CB
#> 9            Sud                       Puglia                       Foggia   FG
#> 10           Sud                     Calabria              Reggio Calabria   RC
#> 11         Isole                      Sicilia                     Siracusa   SR
#> 12         Isole                     Sardegna                      Sassari   SS
#> 13    Nord-ovest                      Liguria                      Imperia   IM
#> 14    Nord-ovest                      Liguria                       Savona   SV
#> 15    Nord-ovest                      Liguria                       Genova   GE
#> 16    Nord-ovest                      Liguria                    La Spezia   SP
#> 17    Nord-ovest                    Lombardia                       Varese   VA
#> 18    Nord-ovest                    Lombardia                         Como   CO
#> 19    Nord-ovest                    Lombardia                      Sondrio   SO
#> 20    Nord-ovest                    Lombardia                       Milano   MI
#> 21    Nord-ovest                    Lombardia                      Bergamo   BG
#> 22    Nord-ovest                    Lombardia                      Brescia   BS
#> 23    Nord-ovest                    Lombardia                        Pavia   PV
#> 24    Nord-ovest                    Lombardia                      Cremona   CR
#> 25    Nord-ovest                    Lombardia                      Mantova   MN
#> 26      Nord-est Trentino-Alto Adige/Südtirol                Bolzano/Bozen   BZ
#> 27      Nord-est Trentino-Alto Adige/Südtirol                       Trento   TN
#> 28      Nord-est                       Veneto                       Verona   VR
#> 29      Nord-est                       Veneto                      Vicenza   VI
#> 30      Nord-est                       Veneto                      Belluno   BL
#> 31      Nord-est                       Veneto                      Treviso   TV
#> 32      Nord-est                       Veneto                      Venezia   VE
#> 33      Nord-est                       Veneto                       Padova   PD
#> 34      Nord-est                       Veneto                       Rovigo   RO
#> 35      Nord-est        Friuli-Venezia Giulia                        Udine   UD
#> 36      Nord-est        Friuli-Venezia Giulia                      Gorizia   GO
#> 37      Nord-est        Friuli-Venezia Giulia                      Trieste   TS
#> 38      Nord-est               Emilia-Romagna                     Piacenza   PC
#> 39      Nord-est               Emilia-Romagna                        Parma   PR
#> 40      Nord-est               Emilia-Romagna           Reggio nell'Emilia   RE
#> 41      Nord-est               Emilia-Romagna                       Modena   MO
#> 42      Nord-est               Emilia-Romagna                      Bologna   BO
#> 43      Nord-est               Emilia-Romagna                      Ferrara   FE
#> 44      Nord-est               Emilia-Romagna                      Ravenna   RA
#> 45      Nord-est               Emilia-Romagna                 Forlì-Cesena   FC
#> 46        Centro                       Marche              Pesaro e Urbino   PU
#> 47        Centro                       Marche                       Ancona   AN
#> 48        Centro                       Marche                     Macerata   MC
#> 49        Centro                       Marche                Ascoli Piceno   AP
#> 50        Centro                      Toscana                Massa-Carrara   MS
#> 51        Centro                      Toscana                        Lucca   LU
#> 52        Centro                      Toscana                      Pistoia   PT
#> 53        Centro                      Toscana                      Firenze   FI
#> 54        Centro                      Toscana                      Livorno   LI
#> 55        Centro                      Toscana                         Pisa   PI
#> 56        Centro                      Toscana                       Arezzo   AR
#> 57        Centro                      Toscana                        Siena   SI
#> 58        Centro                      Toscana                     Grosseto   GR
#> 59        Centro                       Umbria                      Perugia   PG
#> 60        Centro                       Umbria                        Terni   TR
#> 61        Centro                        Lazio                      Viterbo   VT
#> 62        Centro                        Lazio                        Rieti   RI
#> 63        Centro                        Lazio                         Roma   RM
#> 64        Centro                        Lazio                       Latina   LT
#> 65        Centro                        Lazio                    Frosinone   FR
#> 66           Sud                     Campania                      Caserta   CE
#> 67           Sud                     Campania                    Benevento   BN
#> 68           Sud                     Campania                       Napoli   NA
#> 69           Sud                     Campania                     Avellino   AV
#> 70           Sud                     Campania                      Salerno   SA
#> 71           Sud                      Abruzzo                     L'Aquila   AQ
#> 72           Sud                      Abruzzo                       Teramo   TE
#> 73           Sud                      Abruzzo                      Pescara   PE
#> 74           Sud                      Abruzzo                       Chieti   CH
#> 75           Sud                       Puglia                         Bari   BA
#> 76           Sud                       Puglia                      Taranto   TA
#> 77           Sud                       Puglia                     Brindisi   BR
#> 78           Sud                       Puglia                        Lecce   LE
#> 79           Sud                   Basilicata                      Potenza   PZ
#> 80           Sud                   Basilicata                       Matera   MT
#> 81           Sud                     Calabria                      Cosenza   CS
#> 82           Sud                     Calabria                    Catanzaro   CZ
#> 83         Isole                      Sicilia                      Trapani   TP
#> 84         Isole                      Sicilia                      Palermo   PA
#> 85         Isole                      Sicilia                      Messina   ME
#> 86         Isole                      Sicilia                    Agrigento   AG
#> 87         Isole                      Sicilia                Caltanissetta   CL
#> 88         Isole                      Sicilia                         Enna   EN
#> 89         Isole                      Sicilia                      Catania   CT
#> 90         Isole                      Sicilia                       Ragusa   RG
#> 91         Isole                     Sardegna                        Nuoro   NU
#> 92         Isole                     Sardegna                     Cagliari   CA
#> 93      Nord-est        Friuli-Venezia Giulia                    Pordenone   PN
#> 94           Sud                       Molise                      Isernia   IS
#> 95         Isole                     Sardegna                     Oristano   OR
#> 96    Nord-ovest                     Piemonte                       Biella   BI
#> 97    Nord-ovest                    Lombardia                        Lecco   LC
#> 98    Nord-ovest                    Lombardia                         Lodi   LO
#> 99      Nord-est               Emilia-Romagna                       Rimini   RN
#> 100       Centro                      Toscana                        Prato   PO
#> 101          Sud                     Calabria                      Crotone   KR
#> 102          Sud                     Calabria                Vibo Valentia   VV
#> 103   Nord-ovest                     Piemonte         Verbano-Cusio-Ossola   VB
#> 104   Nord-ovest                    Lombardia        Monza e della Brianza   MB
#> 105       Centro                       Marche                        Fermo   FM
#> 106          Sud                       Puglia        Barletta-Andria-Trani   BT
#> 107        Isole                     Sardegna                 Sud Sardegna   SU
#>     code_ripartizione code_regione code_provincia
#> 1                   1            1              1
#> 2                   1            1              2
#> 3                   1            1              3
#> 4                   1            1              4
#> 5                   1            1              5
#> 6                   1            1              6
#> 7                   1            2              7
#> 8                   4           14             70
#> 9                   4           16             71
#> 10                  4           18             80
#> 11                  5           19             89
#> 12                  5           20             90
#> 13                  1            7              8
#> 14                  1            7              9
#> 15                  1            7             10
#> 16                  1            7             11
#> 17                  1            3             12
#> 18                  1            3             13
#> 19                  1            3             14
#> 20                  1            3             15
#> 21                  1            3             16
#> 22                  1            3             17
#> 23                  1            3             18
#> 24                  1            3             19
#> 25                  1            3             20
#> 26                  2            4             21
#> 27                  2            4             22
#> 28                  2            5             23
#> 29                  2            5             24
#> 30                  2            5             25
#> 31                  2            5             26
#> 32                  2            5             27
#> 33                  2            5             28
#> 34                  2            5             29
#> 35                  2            6             30
#> 36                  2            6             31
#> 37                  2            6             32
#> 38                  2            8             33
#> 39                  2            8             34
#> 40                  2            8             35
#> 41                  2            8             36
#> 42                  2            8             37
#> 43                  2            8             38
#> 44                  2            8             39
#> 45                  2            8             40
#> 46                  3           11             41
#> 47                  3           11             42
#> 48                  3           11             43
#> 49                  3           11             44
#> 50                  3            9             45
#> 51                  3            9             46
#> 52                  3            9             47
#> 53                  3            9             48
#> 54                  3            9             49
#> 55                  3            9             50
#> 56                  3            9             51
#> 57                  3            9             52
#> 58                  3            9             53
#> 59                  3           10             54
#> 60                  3           10             55
#> 61                  3           12             56
#> 62                  3           12             57
#> 63                  3           12             58
#> 64                  3           12             59
#> 65                  3           12             60
#> 66                  4           15             61
#> 67                  4           15             62
#> 68                  4           15             63
#> 69                  4           15             64
#> 70                  4           15             65
#> 71                  4           13             66
#> 72                  4           13             67
#> 73                  4           13             68
#> 74                  4           13             69
#> 75                  4           16             72
#> 76                  4           16             73
#> 77                  4           16             74
#> 78                  4           16             75
#> 79                  4           17             76
#> 80                  4           17             77
#> 81                  4           18             78
#> 82                  4           18             79
#> 83                  5           19             81
#> 84                  5           19             82
#> 85                  5           19             83
#> 86                  5           19             84
#> 87                  5           19             85
#> 88                  5           19             86
#> 89                  5           19             87
#> 90                  5           19             88
#> 91                  5           20             91
#> 92                  5           20             92
#> 93                  2            6             93
#> 94                  4           14             94
#> 95                  5           20             95
#> 96                  1            1             96
#> 97                  1            3             97
#> 98                  1            3             98
#> 99                  2            8             99
#> 100                 3            9            100
#> 101                 4           18            101
#> 102                 4           18            102
#> 103                 1            1            103
#> 104                 1            3            108
#> 105                 3           11            109
#> 106                 4           16            110
#> 107                 5           20            111
getNamesIT(unit = "provincia", all_levels = FALSE)
#>   [1] "Torino"                       "Vercelli"                    
#>   [3] "Novara"                       "Cuneo"                       
#>   [5] "Asti"                         "Alessandria"                 
#>   [7] "Valle d'Aosta/Vallée d'Aoste" "Campobasso"                  
#>   [9] "Foggia"                       "Reggio Calabria"             
#>  [11] "Siracusa"                     "Sassari"                     
#>  [13] "Imperia"                      "Savona"                      
#>  [15] "Genova"                       "La Spezia"                   
#>  [17] "Varese"                       "Como"                        
#>  [19] "Sondrio"                      "Milano"                      
#>  [21] "Bergamo"                      "Brescia"                     
#>  [23] "Pavia"                        "Cremona"                     
#>  [25] "Mantova"                      "Bolzano/Bozen"               
#>  [27] "Trento"                       "Verona"                      
#>  [29] "Vicenza"                      "Belluno"                     
#>  [31] "Treviso"                      "Venezia"                     
#>  [33] "Padova"                       "Rovigo"                      
#>  [35] "Udine"                        "Gorizia"                     
#>  [37] "Trieste"                      "Piacenza"                    
#>  [39] "Parma"                        "Reggio nell'Emilia"          
#>  [41] "Modena"                       "Bologna"                     
#>  [43] "Ferrara"                      "Ravenna"                     
#>  [45] "Forlì-Cesena"                 "Pesaro e Urbino"             
#>  [47] "Ancona"                       "Macerata"                    
#>  [49] "Ascoli Piceno"                "Massa-Carrara"               
#>  [51] "Lucca"                        "Pistoia"                     
#>  [53] "Firenze"                      "Livorno"                     
#>  [55] "Pisa"                         "Arezzo"                      
#>  [57] "Siena"                        "Grosseto"                    
#>  [59] "Perugia"                      "Terni"                       
#>  [61] "Viterbo"                      "Rieti"                       
#>  [63] "Roma"                         "Latina"                      
#>  [65] "Frosinone"                    "Caserta"                     
#>  [67] "Benevento"                    "Napoli"                      
#>  [69] "Avellino"                     "Salerno"                     
#>  [71] "L'Aquila"                     "Teramo"                      
#>  [73] "Pescara"                      "Chieti"                      
#>  [75] "Bari"                         "Taranto"                     
#>  [77] "Brindisi"                     "Lecce"                       
#>  [79] "Potenza"                      "Matera"                      
#>  [81] "Cosenza"                      "Catanzaro"                   
#>  [83] "Trapani"                      "Palermo"                     
#>  [85] "Messina"                      "Agrigento"                   
#>  [87] "Caltanissetta"                "Enna"                        
#>  [89] "Catania"                      "Ragusa"                      
#>  [91] "Nuoro"                        "Cagliari"                    
#>  [93] "Pordenone"                    "Isernia"                     
#>  [95] "Oristano"                     "Biella"                      
#>  [97] "Lecco"                        "Lodi"                        
#>  [99] "Rimini"                       "Prato"                       
#> [101] "Crotone"                      "Vibo Valentia"               
#> [103] "Verbano-Cusio-Ossola"         "Monza e della Brianza"       
#> [105] "Fermo"                        "Barletta-Andria-Trani"       
#> [107] "Sud Sardegna"