Static maps for United Kingdom

Function to produce static maps for United Kingdom statistical unit.

Usage

mappingUK(data, var = NULL, colID = NULL, 
          type = c("static", "interactive"), 
          typeStatic = c("tmap", "choro.cart", "typo", "bar"),
          unit = c("country", "county"), year = c("2020", "2019"), 
          matchWith = c("name", "code"), scale = c("500", "20"), 
          dir = NULL, add_text = NULL, subset = NULL, 
          facets = NULL, aggregation_fun = sum, aggregation_unit = NULL, 
          options = mapping.options())

Arguments

data

a data.frame object with variables to display or a UK object produced by UK function. If object of class UK, arguments unit, year, and matchWith will be ignored

var

character value(s) or columns number(s) indicating the variable to plot

colID

character value or columns number indicating the column with unit names

type

if generates static or interactive map

typeStatic

type of static map

unit

the type of Italian statistical unit

year

year of the unit

matchWith

the type of id to check:

"name"	if unit names).

"code"	if unit code

scale

the scale of a map

dir

local directory in which shape files are stored

add_text

character name indicating the column with text labels

subset

a formula indicating the condition to subset the data. See the details section

facets

variable(s) name to split the data

aggregation_fun

function to use when data are aggregated

aggregation_unit

variable name by which the unit are aggregated

options

a list with options using mapping.options function

Details

If data is a object of class "UK" generated using the UK function, the arguments unit, and year are ignored, because the object already contains the coordinates.

The aggregation_unit provides an aggregation for a user specified variable in data, or for larger statistical unit, automatically provided when the function link the data with the coordinates. For example, if data are of type county, we will have variables for larger aggregate unit, that is country variables. Look at UK for more details.

subset provide an expression to subsetting the data using a formula, with the logical operators. For example data can be subsetting as follows: ~I("Variable 1" == "condition 1" & "Variable 2" != "condition 2") or for example, ~I("Variable 1" > "condition 1" | "Variable 2" != "condition 2").

Value

Return a map. For tmap type, the function also returns a tmap object.

References

Giraud, T. and Lambert, N. (2016). cartography: Create and Integrate Maps in your R Workflow. JOSS, 1(4). doi: 10.21105/joss.00054.

Pebesma, E., 2018. Simple Features for R: Standardized Support for Spatial Vector Data. The R Journal 10 (1), 439-446, https://doi.org/10.32614/RJ-2018-009

Tennekes M (2018). “tmap: Thematic Maps in R.” _Journalstatisticaltical Software_, *84*(6), 1-39. doi: 10.18637/jss.v084.i06 (URL: https://doi.org/10.18637/jss.v084.i06).

See also

Examples

data("popUK")

uk <- UK(data = popUK, unit = "county", matchWith = "code", check.unit.names = FALSE)